Refactoring after second code review
