+++ sip-test-kit/src/main/java/de/ikor/sip/foundation/testkit/workflow/whenphase/routeinvoker/impl/RestRouteInvoker.java
@@ -1,6 +1,5 @@
 package de.ikor.sip.foundation.testkit.workflow.whenphase.routeinvoker.impl;
 
-import de.ikor.sip.foundation.testkit.util.SIPEndpointResolver;
 import de.ikor.sip.foundation.testkit.util.SIPExchangeHelper;
 import de.ikor.sip.foundation.testkit.workflow.whenphase.routeinvoker.RouteInvoker;
 import java.util.Optional;
@@ -26,9 +25,7 @@ public class RestRouteInvoker implements RouteInvoker {
 
   @Override
   public Optional<Exchange> invoke(Exchange inputExchange) {
-    Endpoint endpoint =
-        SIPEndpointResolver.resolveEndpoint(
-            SIPExchangeHelper.getRouteId(inputExchange), camelContext);
+    Endpoint endpoint = SIPExchangeHelper.resolveEndpoint(inputExchange, camelContext);
     Exchange exchange =
         producerTemplate.send(extractRESTEndpointURI((RestEndpoint) endpoint), inputExchange);
     return Optional.of(exchange);
